Excel Project
This project offers an in-depth analysis of a dataset that details information about bike buyers. The process includes data cleaning, transformation, the creation of pivot tables, and visualization of the data. The final product is a dynamic dashboard with slicers, allowing users to filter data by marital status, region, and education level.